AgentDeploy/requirements_agents.txt
# Azure Cost Analysis Agents - Requirements
# 
# Install with: pip install -r requirements_agents.txt

# Microsoft Agent Framework (Preview)
agent-framework-azure-ai

# HTTP client for Function App communication
aiohttp>=3.9.0

# Azure authentication
azure-identity>=1.15.0

# Optional: Enhanced logging and debugging
rich>=13.0.0

# Optional: Environment variable management
python-dotenv>=1.0.0